app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets>=11.0
pydantic
pydantic-settings
python-dotenv
httpx
mem0ai
tavily

# Single persistence layer
qdrant-client>=1.9.0

# Encrypt agent API keys at rest
cryptography>=41.0.0

# Note: Mantle packages removed as they're not used in the backend
# The code uses httpx directly for Mantle RPC calls

